Style the views using Bootstrap
The sample app in the book uses the bootstrap-sass gem. Since that
is quite old and only for bootstrap 3, I am using yarn to add
bootstrap to the project. Procedure:

    $yarn add bootstrap jquery popper.js

In order for bootstrap to correctly use sass

    $cd app/assets/stylesheets
    $mv application.css application.scss

In application.scss, add

    @import "bootstrap/scss/bootstrap"

There are other things to be done in order to import bootstrap
components, but I haven't done any of that yet.