Configure and add Devise User model
A devise User model is generated and the devise views (from gem) are
added to the application.

    $rails generate devise User
    $rails generate devise:views

Among the views, the passwords, registrations and sessions form are
styled using Bootstrap classes (just form-group, form-control, and btn
nothing fancy).

To the user model, additional attributes (first_name and last_name) are
added. Fields for these attributes are also added to the registrations
forms. Further, a migration is generated for these attributes on the
User model. Finally, code is added to the ApplicationController to permit
these additional attributes as strong parameters.

In the application controller, ensure that user is authenticated before
any controller action is performed. Except for the HomeController, which
is the application root (landing page). All this is done as per the
"Getting Started" instruction in devise README doc.