Add Note model using scaffold
Scaffold generated using

    $rails g scaffold Note message:text project:references user:references

Association:
    Note belongs to user, project
    Both user and project have many notes

Callbacks are created in ApplicationController to ensure that the
notes resource is only available to the project owner. The resource
is namespaced under user.

The scaffold adds user_id and project_id fields to the generated
forms. These should be removed and the fields need to be filled in
server side using associations.

note partial is also created to show related notes in the project
view.

A better scaffolding command would have been to use user/note as
that would have autogenerated views with the correct links. Check
the output of `$rails routes` to get the correct paths.